ML_image_recognition/requirements.txt
tensorflow>=2.3.0
numpy>=1.18.5
pandas>=1.0.0
Pillow>=7.0.0
matplotlib>=3.2.0
scikit-learn>=0.22.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
opencv-python>=4.5.0
beautifulsoup4>=4.9.0
requests>=2.25.0
flask>=2.0.0
flask-cors>=4.0.0